+++ src/Symfony/Bundle/WebProfilerBundle/Controller/ExceptionController.php
@@ -11,7 +11,7 @@
 
 namespace Symfony\Bundle\WebProfilerBundle\Controller;
 
-use Symfony\Component\Debug\ExceptionHandler;
+use Symfony\Component\ErrorHandler\ErrorRenderer\HtmlErrorRenderer;
 use Symfony\Component\HttpFoundation\Response;
 use Symfony\Component\HttpKernel\Debug\FileLinkFormatter;
 use Symfony\Component\HttpKernel\Exception\NotFoundHttpException;
@@ -30,14 +30,18 @@ class ExceptionController
     protected $twig;
     protected $debug;
     protected $profiler;
-    private $fileLinkFormat;
+    private $errorRenderer;
 
-    public function __construct(Profiler $profiler = null, Environment $twig, bool $debug, FileLinkFormatter $fileLinkFormat = null)
+    public function __construct(Profiler $profiler = null, Environment $twig, bool $debug, FileLinkFormatter $fileLinkFormat = null, HtmlErrorRenderer $errorRenderer = null)
     {
         $this->profiler = $profiler;
         $this->twig = $twig;
         $this->debug = $debug;
-        $this->fileLinkFormat = $fileLinkFormat;
+        $this->errorRenderer = $errorRenderer;
+
+        if (null === $errorRenderer) {
+            $this->errorRenderer = new HtmlErrorRenderer($debug, $this->twig->getCharset(), $fileLinkFormat);
+        }
     }
 
     /**
@@ -61,9 +65,7 @@ class ExceptionController
         $template = $this->getTemplate();
 
         if (!$this->twig->getLoader()->exists($template)) {
-            $handler = new ExceptionHandler($this->debug, $this->twig->getCharset(), $this->fileLinkFormat);
-
-            return new Response($handler->getContent($exception), 200, ['Content-Type' => 'text/html']);
+            return new Response($this->errorRenderer->getBody($exception), 200, ['Content-Type' => 'text/html']);
         }
 
         $code = $exception->getStatusCode();
@@ -97,13 +99,10 @@ class ExceptionController
 
         $this->profiler->disable();
 
-        $exception = $this->profiler->loadProfile($token)->getCollector('exception')->getException();
         $template = $this->getTemplate();
 
         if (!$this->templateExists($template)) {
-            $handler = new ExceptionHandler($this->debug, $this->twig->getCharset(), $this->fileLinkFormat);
-
-            return new Response($handler->getStylesheet($exception), 200, ['Content-Type' => 'text/css']);
+            return new Response($this->errorRenderer->getStylesheet(), 200, ['Content-Type' => 'text/css']);
         }
 
         return new Response($this->twig->render('@WebProfiler/Collector/exception.css.twig'), 200, ['Content-Type' => 'text/css']);