+++ src/Symfony/Component/Debug/Exception/FatalThrowableError.php
@@ -11,41 +11,13 @@
 
 namespace Symfony\Component\Debug\Exception;
 
+use Symfony\Component\ErrorHandler\Exception\FatalThrowableError as BaseFatalThrowableError;
+
+@trigger_error(sprintf('The "%s" class is deprecated since Symfony 4.4, use "%s" instead.', FatalThrowableError::class, BaseFatalThrowableError::class), E_USER_DEPRECATED);
+
 /**
- * Fatal Throwable Error.
- *
- * @author Nicolas Grekas <p@tchwork.com>
+ * @deprecated since Symfony 4.4, use Symfony\Component\ErrorHandler\Exception\FatalThrowableError instead.
  */
-class FatalThrowableError extends FatalErrorException
+class FatalThrowableError extends BaseFatalThrowableError
 {
-    private $originalClassName;
-
-    public function __construct(\Throwable $e)
-    {
-        $this->originalClassName = \get_class($e);
-
-        if ($e instanceof \ParseError) {
-            $severity = E_PARSE;
-        } elseif ($e instanceof \TypeError) {
-            $severity = E_RECOVERABLE_ERROR;
-        } else {
-            $severity = E_ERROR;
-        }
-
-        \ErrorException::__construct(
-            $e->getMessage(),
-            $e->getCode(),
-            $severity,
-            $e->getFile(),
-            $e->getLine(),
-            $e->getPrevious()
-        );
-
-        $this->setTrace($e->getTrace());
-    }
-
-    public function getOriginalClassName(): string
-    {
-        return $this->originalClassName;
-    }
 }