+++ src/Symfony/Component/Debug/CHANGELOG.md
@@ -1,6 +1,18 @@
 CHANGELOG
 =========
 
+4.4.0
+-----
+
+* deprecated the `BufferingLogger`, `ErrorHandler` and `ExceptionHandler` classes, 
+  they have been moved to the `ErrorHandler` component
+* deprecated the `FatalErrorHandlerInterface`, `ClassNotFoundFatalErrorHandler`,  
+  `UndefinedFunctionFatalErrorHandler` and `UndefinedMethodFatalErrorHandler` classes,
+  they have been moved to the `ErrorHandler` component
+* deprecated the `ClassNotFoundException`, `FatalErrorException`, `FatalThrowableError`,
+  `FlattenException`, `OutOfMemoryException`, `SilencedErrorContext`, `UndefinedFunctionException`,
+  and `UndefinedMethodException`, they have been moved to the `ErrorHandler` component
+
 4.3.0
 -----